Final touch, all the new endpoint are now in an extension namespace
