+++ src/main/extension.js
@@ -1,0 +1,148 @@
+import Stream from 'stream'
+import * as transformers from './transformers'
+
+import { isValidPrefix, isValidBucketName, uriEscape,
+  isBoolean, isNumber, isString,
+  pipesetup } from './helpers.js'
+  
+
+export default class extension {
+
+  constructor(client)
+  {
+    this.client = client;
+  }
+
+  // List the objects in the bucket using S3 ListObjects V2 With Metadata
+  //
+  // __Arguments__
+  // * `bucketName` _string_: name of the bucket
+  // * `prefix` _string_: the prefix of the objects that should be listed (optional, default `''`)
+  // * `recursive` _bool_: `true` indicates recursive style listing and `false` indicates directory style listing delimited by '/'. (optional, default `false`)
+  // * `startAfter` _string_: Specifies the key to start after when listing objects in a bucket. (optional, default `''`)
+  //
+  // __Return Value__
+  // * `stream` _Stream_: stream emitting the objects in the bucket, the object is of the format:
+  //   * `obj.name` _string_: name of the object
+  //   * `obj.prefix` _string_: name of the object prefix
+  //   * `obj.size` _number_: size of the object
+  //   * `obj.etag` _string_: etag of the object
+  //   * `obj.lastModified` _Date_: modified time stamp
+
+  listObjectsV2WithMetadata(bucketName, prefix, recursive, startAfter) {
+    if (prefix === undefined) prefix = ''
+    if (recursive === undefined) recursive = false
+    if (startAfter === undefined) startAfter = ''
+    if (!isValidBucketName(bucketName)) {
+      throw new errors.InvalidBucketNameError('Invalid bucket name: ' + bucketName)
+    }
+    if (!isValidPrefix(prefix)) {
+      throw new errors.InvalidPrefixError(`Invalid prefix : ${prefix}`)
+    }
+    if (!isString(prefix)) {
+      throw new TypeError('prefix should be of type "string"')
+    }
+    if (!isBoolean(recursive)) {
+      throw new TypeError('recursive should be of type "boolean"')
+    }
+    if (!isString(startAfter)) {
+      throw new TypeError('startAfter should be of type "string"')
+    }
+    // if recursive is false set delimiter to '/'
+    var delimiter = recursive ? '' : '/'
+    var continuationToken = ''
+    var objects = []
+    var ended = false
+    var readStream = Stream.Readable({objectMode: true})
+    readStream._read = () => {
+      // push one object per _read()
+      if (objects.length) {
+        readStream.push(objects.shift())
+        return
+      }
+      if (ended) return readStream.push(null)
+      // if there are no objects to push do query for the next batch of objects
+      this.listObjectsV2WithMetadataQuery(bucketName, prefix, continuationToken, delimiter, 1000, startAfter)
+        .on('error', e => readStream.emit('error', e))
+        .on('data', result => {
+          if (result.isTruncated) {
+            continuationToken = result.nextContinuationToken
+          } else {
+            ended = true
+          }
+          objects = result.objects
+          readStream._read()
+        })
+    }
+    return readStream
+  }
+
+  // listObjectsV2WithMetadataQuery - (List Objects V2 with metadata) - List some or all (up to 1000) of the objects in a bucket.
+  //
+  // You can use the request parameters as selection criteria to return a subset of the objects in a bucket.
+  // request parameters :-
+  // * `bucketName` _string_: name of the bucket
+  // * `prefix` _string_: Limits the response to keys that begin with the specified prefix.
+  // * `continuation-token` _string_: Used to continue iterating over a set of objects.
+  // * `delimiter` _string_: A delimiter is a character you use to group keys.
+  // * `max-keys` _number_: Sets the maximum number of keys returned in the response body.
+  // * `start-after` _string_: Specifies the key to start after when listing objects in a bucket.
+  
+  listObjectsV2WithMetadataQuery(bucketName, prefix, continuationToken, delimiter, maxKeys, startAfter) {
+    if (!isValidBucketName(bucketName)) {
+      throw new errors.InvalidBucketNameError('Invalid bucket name: ' + bucketName)
+    }
+    if (!isString(prefix)) {
+      throw new TypeError('prefix should be of type "string"')
+    }
+    if (!isString(continuationToken)) {
+      throw new TypeError('continuationToken should be of type "string"')
+    }
+    if (!isString(delimiter)) {
+      throw new TypeError('delimiter should be of type "string"')
+    }
+    if (!isNumber(maxKeys)) {
+      throw new TypeError('maxKeys should be of type "number"')
+    }
+    if (!isString(startAfter)) {
+      throw new TypeError('startAfter should be of type "string"')
+    }
+    var queries = []
+
+    // Call for listing objects v2 API
+    queries.push(`list-type=2`)
+    // escape every value in query string, except maxKeys
+    queries.push(`prefix=${uriEscape(prefix)}`)
+    queries.push(`delimiter=${uriEscape(delimiter)}`)
+    queries.push(`metadata=true`)
+
+    if (continuationToken) {
+      continuationToken = uriEscape(continuationToken)
+      queries.push(`continuation-token=${continuationToken}`)
+    }
+    // Set start-after
+    if (startAfter) {
+      startAfter = uriEscape(startAfter)
+      queries.push(`start-after=${startAfter}`)
+    }
+    // no need to escape maxKeys
+    if (maxKeys) {
+      if (maxKeys >= 1000) {
+        maxKeys = 1000
+      }
+      queries.push(`max-keys=${maxKeys}`)
+    }
+    queries.sort()
+    var query = ''
+    if (queries.length > 0) {
+      query = `${queries.join('&')}`
+    }
+    var method = 'GET'
+    var transformer = transformers.getListObjectsV2WithMetadataTransformer()
+    this.client.makeRequest({method, bucketName, query}, '', 200, '', true, (e, response) => {
+      if (e) return transformer.emit('error', e)
+      pipesetup(response, transformer)
+    })
+    return transformer
+  }
+}
